+++ Content.Client/Physics/Controllers/ConveyorController.cs
@@ -1,0 +1,8 @@
+﻿using Content.Shared.Physics.Controllers;
+
+namespace Content.Client.Physics.Controllers;
+
+public sealed class ConveyorController : SharedConveyorController
+{
+    //Class is empty, needed for prediction and networking
+}
