+++ Content.Server/Power/Components/ApcPowerReceiverComponent.cs
@@ -70,6 +70,7 @@ namespace Content.Server.Power.Components
 
     /// <summary>
     /// Raised whenever an ApcPowerReceiver becomes powered / unpowered.
+    /// Does nothing on the client.
     /// </summary>
     [ByRefEvent]
     public readonly record struct PowerChangedEvent(bool Powered, float ReceivingPower)
@@ -77,4 +78,5 @@ namespace Content.Server.Power.Components
         public readonly bool Powered = Powered;
         public readonly float ReceivingPower = ReceivingPower;
     }
+
 }
