+++ Content.Shared/Conveyor/ActiveConveyorComponent.cs
@@ -1,4 +1,6 @@
-namespace Content.Server.Conveyor;
+using Robust.Shared.GameStates;
+
+namespace Content.Shared.Conveyor;
 
 /// <summary>
 /// Used to track which conveyors are relevant in case there's a lot of them.
